// src/pages/products/types/form.types.ts
export interface ProductFormState {
  name: string;
  parentCategoryId: string;
  categoryLevel2Id: string;
  categoryLevel3Id: string;
  categoryLevel4Id: string;
  barcode: string;
  stockQuantity: string;
  stockUnit: string;
  salePrice: string;
  purchasePrice: string;
  imageUrl: string;
  isActive: boolean;
  description: string;
  companyId: string;
}